+++ src/retry-interceptor.ts
@@ -1,7 +1,13 @@
 import { PLATFORM } from 'aurelia-pal';
 import { Interceptor, RetryConfiguration } from './interfaces';
+import { HttpClient } from './http-client';
 
-export const retryStrategy = {
+export const retryStrategy: {
+  fixed: 0;
+  incremental: 1;
+  exponential: 2;
+  random: 3;
+} = {
   fixed: 0,
   incremental: 1,
   exponential: 2,
@@ -26,26 +32,27 @@ export class RetryInterceptor implements Interceptor {
     }
   }
 
-  request(request) {
-    if (!request.retryConfig) {
-      request.retryConfig = Object.assign({}, this.retryConfig);
-      request.retryConfig.counter = 0;
+  request(request: Request) {
+    let $r = request as Request & { retryConfig?: RetryConfiguration };
+    if (!$r.retryConfig) {
+      $r.retryConfig = Object.assign({}, this.retryConfig);
+      $r.retryConfig.counter = 0;
     }
 
     // do this on every request
-    request.retryConfig.requestClone = request.clone();
+    $r.retryConfig.requestClone = request.clone();
 
     return request;
   }
 
-  response(response, request) {
+  response(response: Response, request: Request) {
     // retry was successful, so clean up after ourselves
-    delete request.retryConfig;
+    delete (request as any).retryConfig;
     return response;
   }
 
-  responseError(error, request, httpClient) {
-    const { retryConfig } = request;
+  responseError(error: Response, request: Request, httpClient: HttpClient) {
+    const { retryConfig } = request as Request & { retryConfig: RetryConfiguration };
     const { requestClone } = retryConfig;
     return Promise.resolve().then(() => {
       if (retryConfig.counter < retryConfig.maxRetries) {
@@ -66,12 +73,12 @@ export class RetryInterceptor implements Interceptor {
           }
 
           // no more retries, so clean up
-          delete request.retryConfig;
+          delete (request as any).retryConfig;
           throw error;
         });
       }
       // no more retries, so clean up
-      delete request.retryConfig;
+      delete (request as any).retryConfig;
       throw error;
     });
   }
@@ -81,7 +88,7 @@ function calculateDelay(retryConfig: RetryConfiguration) {
   const { interval, strategy, minRandomInterval, maxRandomInterval, counter } = retryConfig;
 
   if (typeof (strategy) === 'function') {
-    return retryConfig.strategy(counter);
+    return (retryConfig.strategy as Function)(counter);
   }
 
   switch (strategy) {
@@ -112,4 +119,9 @@ const retryStrategies = [
   (retryCount, interval, minRandomInterval = 0, maxRandomInterval = 60000) => {
     return Math.random() * (maxRandomInterval - minRandomInterval) + minRandomInterval;
   }
+] as [
+  (interval: number) => number,
+  (retryCount: number, interval: number) => number,
+  (retryCount: number, interval: number) => number,
+  (retryCount: number, interval: number, minRandomInterval?: number, maxRandomInterval?: number) => number
 ];
